Move the Windows comparison tests to their own package
WmiComparisonTest, PerfmonComparisonTest and RegistryComparisonTest are
Windows-only, but sat in oshi.comparison, which no path filter selects. The
AppVeyor filter looks for a windows path segment, so a push touching only these
files ran nothing: that is how a commit whose comparison tests did not compile
reached CI green.

Put them under oshi.comparison.windows so the existing filters select them by
construction rather than by growing a list. ComparisonAssertions becomes public
to stay reachable from the subpackage; it is test-only code in a module that
publishes nothing.
