+++ oshi-benchmark/src/test/java/oshi/comparison/windows/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Comparison tests for the Windows-only query mechanisms: WMI, performance counters, and the registry. They live in
+ * their own package so that CI path filters keyed on a {@code windows} path segment select them.
+ */
+package oshi.comparison.windows;
